Prepare code to enable float-conversion warning, never use abs(…) without std:: due to this mess: https://stackoverflow.com/questions/21392627/abs-vs-stdabs-what-does-the-reference-say

+++ gui/qt4/OpenGLManager.cpp
@@ -60,10 +60,10 @@ void OpenGLManager::startTimerSlot(){
 	startTimer(50);
 }
 
-int OpenGLManager::waitForNewView(float timeout,bool center){
+int OpenGLManager::waitForNewView(double timeout,bool center){
 	size_t origViewCount=views.size();
 	emitCreateView();
-	float t=0;
+	double t=0;
 	while(views.size()!=origViewCount+1){
 		usleep(50000); t+=.05;
 		// wait at most 5 secs
