+++ py/wrapper/yadeWrapper.cpp
@@ -296,7 +296,7 @@ class pyBodyContainer{
 			}
 			
 			//get pointer lists of spheres, that should be replaced:
-			int numReplaceTmp = round(num*amounts[ii]);
+			int numReplaceTmp = int(std::round(num*amounts[ii]));
 			vector<shared_ptr<Body> > bpListTmp(numReplaceTmp);
 			int a = 0, c = 0;//counters
 			vector<int> posTmp;
